refactor : LottoBoughtTicket 제거
상속을 함으로써 얻는 이익은 보이지 않고 로직만 복잡하게 만들기 때문에 LottoBoughtTicket 클래스는 제거
