+++ src/main/java/lotto/domain/lottos/LottoTicket.java
@@ -1,5 +1,8 @@
 package lotto.domain.lottos;
 
+import lotto.domain.lottos.rank.LottoRank;
+import lotto.domain.lottos.winnerlotto.LottoWinner;
+
 import java.util.*;
 
 public class LottoTicket {
@@ -16,6 +19,19 @@ public class LottoTicket {
         Objects.requireNonNull(lottoNumbers, NULL_ERROR_MESSAGE);
         validateLottoTicket(lottoNumbers);
         this.lottoNumbers = lottoNumbers;
+    }
+
+    public LottoRank getRank(final LottoWinner lottoWinner) {
+        int matchCount = calculateMatchCount(lottoWinner.getLottoWinnerTicket());
+        boolean matchBonusNumber = this.isContain(lottoWinner.getLottoWinnerBonusNumber());
+        return LottoRank.matchLottoRank(matchCount, matchBonusNumber);
+    }
+
+    private int calculateMatchCount(final LottoTicket lottoWinnerTicket) {
+        return (int) lottoWinnerTicket.getLottoNumbers()
+                .stream()
+                .filter(this::isContain)
+                .count();
     }
 
     public boolean isContain(final LottoNumber lottoNumber) {
