+++ src/main/java/lotto/domain/lottos/LottoTickets.java
@@ -24,7 +24,7 @@ public class LottoTickets {
 
     public void putLottoResult(Map<LottoRank, Integer> lottoResult, LottoWinner lottoWinner) {
         this.lottoTickets.forEach(lottoTicket -> {
-            LottoBoughtTicket lottoBoughtTicket = (LottoBoughtTicket) lottoTicket;
+            LottoTicket lottoBoughtTicket = lottoTicket;
             LottoRank rank = lottoBoughtTicket.getRank(lottoWinner);
             lottoResult.put(rank, lottoResult.getOrDefault(rank, 0) + INCREMENT_COUNT);
         });
